Feat/select sensor (#30)
* select sensor for cpu and gpu

* brightness icons

* new dropdown layout

+++ core/common/src/main/kotlin/hwinfo/SensorElement.kt
@@ -4,8 +4,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class SensorElement(
-    val dwSensorId: Int,
-    val dwSensorInst: Int,
+    val dwSensorId: Long,
+    val dwSensorInst: Long,
     val szSensorNameOrig: String,
     val szSensorNameUser: String,
 )
